WIP: reoganisation of Managed Item UI and additional deployment options
