+++ src/Certify.UI/App.xaml
@@ -21,11 +21,9 @@
             <Style TargetType="TextBlock" x:Key="Subheading">
                 <Setter Property="HorizontalAlignment" Value="Left" />
                 <Setter Property="FontFamily" Value="Segoe UI Semibold" />
-                <Setter Property="FontSize" Value="14" />
-                <Setter Property="Padding" Value="4" />
+                <Setter Property="FontSize" Value="16" />
+                <Setter Property="Padding" Value="0,4,0,4" />
                 <Setter Property="HorizontalAlignment" Value="Stretch" />
-
-                <Setter Property="Background" Value="{DynamicResource {x:Static SystemColors.ControlBrushKey}}" />
             </Style>
             <Style TargetType="{x:Type Controls:WindowButtonCommands}"
                    BasedOn="{StaticResource MahApps.Metro.Styles.WindowButtonCommands.Win10}" />
