+++ src/Certify.UI/Controls/ManagedItem/Deployment.xaml
@@ -1,0 +1,52 @@
+﻿<UserControl x:Class="Certify.UI.Controls.ManagedItem.Deployment"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:local="clr-namespace:Certify.UI.Controls.ManagedItem"
+             xmlns:Resources="clr-namespace:Certify.Locales;assembly=Certify.Locales"
+              xmlns:utils="clr-namespace:Certify.UI.Utils"
+             mc:Ignorable="d" Background="{DynamicResource {x:Static SystemColors.ControlLightLightBrushKey}}">
+    <UserControl.Resources>
+        <BooleanToVisibilityConverter x:Key="BoolToVisConverter" />
+        <utils:InverseBooleanConverter x:Key="InvBoolConverter" />
+        <utils:OptionalBooleanToVisibilityConverter True="Collapsed" False="Visible" x:Key="InvBoolVisConverter" />
+        <utils:NullVisibilityConverter x:Key="NullCollapsedConverter" />
+        <utils:ManagedItemHealthVisibilityConverter x:Key="HealthVisibleConverter" />
+        <utils:NullVisibilityConverter x:Key="NullVisibleConverter" Null="Visible" NotNull="Collapsed" />
+    </UserControl.Resources>
+    <DockPanel LastChildFill="True" Margin="4,0,0,0">
+        <TextBlock DockPanel.Dock="Top" Style="{StaticResource Subheading}">Certificate Deployment</TextBlock>
+
+        <TextBlock DockPanel.Dock="Top" Text="Choose how to apply the certificate:" Margin="0,0,0,8" />
+        <StackPanel DockPanel.Dock="Top" Margin="8,0,0,0">
+            <RadioButton GroupName="DeploymentType" Margin="0,0,0,4" IsChecked="True" Content="Apply to matching hostname bindings on selected IIS site (default)." />
+            <RadioButton GroupName="DeploymentType" Margin="0,0,0,4" Content="Apply to all bindings on selected IIS site." />
+            <RadioButton GroupName="DeploymentType" Margin="0,0,0,4" Content="Apply to matching hostname bindings on all IIS sites." HorizontalAlignment="Left" Width="712" />
+            <RadioButton GroupName="DeploymentType" Margin="0,0,0,4" Content="Apply to matching domain bindings, all managed sites." />
+            <RadioButton GroupName="DeploymentType" Margin="0,0,0,4" Content="No deployment." />
+        </StackPanel>
+        <TextBlock DockPanel.Dock="Top" Style="{StaticResource Subheading}">IIS Binding Settings</TextBlock>
+
+        <StackPanel DockPanel.Dock="Top" Orientation="Vertical" Margin="4,16,0,0">
+            <RadioButton Margin="0,0,0,4" IsChecked="{Binding SelectedItem.RequestConfig.PerformAutomatedCertBinding}" GroupName="BindingType" Content="{x:Static Resources:SR.ManagedItemSettings_AutoUpdateBinding}" />
+            <RadioButton Margin="0,0,0,4" IsChecked="{Binding SelectedItem.RequestConfig.PerformAutomatedCertBinding, Mode=OneWay, Converter={StaticResource ResourceKey=InvBoolConverter}}" GroupName="BindingType" Content="{x:Static Resources:SR.ManagedItemSettings_UseSpecificBinding}" />
+        </StackPanel>
+
+        <StackPanel  DockPanel.Dock="Top" Orientation="Vertical" IsEnabled="{Binding SelectedItem.RequestConfig.PerformAutomatedCertBinding, Converter={StaticResource ResourceKey=InvBoolConverter}}" Margin="32,8,0,0">
+            <StackPanel Orientation="Horizontal">
+                <Label Width="131" Content="{x:Static Resources:SR.ManagedItemSettings_BindIP}" />
+                <ComboBox ItemsSource="{Binding HostIPAddresses}" DisplayMemberPath="Description" SelectedValue="{Binding SelectedItem.RequestConfig.BindingIPAddress}" SelectedValuePath="IPAddress" Width="225" />
+            </StackPanel>
+            <StackPanel Orientation="Horizontal">
+                <Label Width="131" Content="{x:Static Resources:SR.ManagedItemSettings_BindPort}" />
+                <TextBox Text="{Binding SelectedItem.RequestConfig.BindingPort}" Width="225" />
+            </StackPanel>
+            <StackPanel Orientation="Horizontal">
+                <Label Width="131" Content="{x:Static Resources:SR.ManagedItemSettings_UseSNI}" />
+                <CheckBox IsChecked="{Binding SelectedItem.RequestConfig.BindingUseSNI}"></CheckBox>
+            </StackPanel>
+            <Label Content="{x:Static Resources:SR.ManagedItemSettings_BindingNote}" />
+        </StackPanel>
+    </DockPanel>
+</UserControl>
