+++ src/Certify.UI/Controls/ManagedItem/Deployment.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Certify.UI.Controls.ManagedItem
+{
+    /// <summary>
+    /// Interaction logic for Deployment.xaml 
+    /// </summary>
+    public partial class Deployment : UserControl
+    {
+        public Deployment()
+        {
+            InitializeComponent();
+        }
+    }
+}
