+++ src/Certify.UI/Controls/ManagedItem/MiscOptions.xaml
@@ -1,0 +1,74 @@
+﻿<UserControl x:Class="Certify.UI.Controls.ManagedItem.MiscOptions"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:local="clr-namespace:Certify.UI.Controls.ManagedItem"
+             xmlns:Resources="clr-namespace:Certify.Locales;assembly=Certify.Locales"
+              xmlns:utils="clr-namespace:Certify.UI.Utils"
+
+             mc:Ignorable="d" Background="{DynamicResource {x:Static SystemColors.ControlLightLightBrushKey}}">
+    <UserControl.Resources>
+        <BooleanToVisibilityConverter x:Key="BoolToVisConverter" />
+        <utils:InverseBooleanConverter x:Key="InvBoolConverter" />
+        <utils:OptionalBooleanToVisibilityConverter True="Collapsed" False="Visible" x:Key="InvBoolVisConverter" />
+        <utils:NullVisibilityConverter x:Key="NullCollapsedConverter" />
+        <utils:ManagedItemHealthVisibilityConverter x:Key="HealthVisibleConverter" />
+        <utils:NullVisibilityConverter x:Key="NullVisibleConverter" Null="Visible" NotNull="Collapsed" />
+    </UserControl.Resources>
+    <DockPanel LastChildFill="True" Margin="8,8,8,8">
+        <TextBlock DockPanel.Dock="Top" Style="{StaticResource Subheading}">General Options</TextBlock>
+        <StackPanel DockPanel.Dock="Top" x:Name="GeneralOptions" Margin="2,0,0,0" Background="{DynamicResource {x:Static SystemColors.ControlLightLightBrushKey}}">
+
+            <!-- General Options-->
+            <StackPanel DockPanel.Dock="Top" Orientation="Horizontal" Margin="4,8,0,4">
+                <CheckBox Content="{x:Static Resources:SR.ManagedItemSettings_EnableAutoRenewal}" IsChecked="{Binding SelectedItem.IncludeInAutoRenew}" Margin="136,0,8,0" />
+                <CheckBox Content="{x:Static Resources:SR.ManagedItemSettings_NotifyPrimaryContactOnRenewalFailure}" IsChecked="{Binding SelectedItem.RequestConfig.EnableFailureNotifications}" />
+            </StackPanel>
+        </StackPanel>
+        <DockPanel DockPanel.Dock="Top" Margin="0,0,0,12">
+            <Label Content="{x:Static Resources:SR.Comments}" />
+            <TextBox Height="50">
+                <TextBox.Text>
+                    <Binding  Path="SelectedItem.Comments" UpdateSourceTrigger="PropertyChanged" />
+                </TextBox.Text>
+            </TextBox>
+        </DockPanel>
+        <TextBlock DockPanel.Dock="Top" Style="{StaticResource Subheading}">Actions</TextBlock>
+
+        <StackPanel Orientation="Vertical" Margin="12">
+
+            <Button x:Name="OpenLogFile" Content="{x:Static Resources:SR.ManagedItemSettings_OpenLogFile}" Click="OpenLogFile_Click" Width="250" HorizontalAlignment="Right" />
+            <Label x:Name="CertPath" Content="{Binding SelectedItem.CertificatePath, TargetNullValue={x:Static Resources:SR.ManagedItemSettings_CertificatePathEmpty}}" />
+            <Button x:Name="OpenCertificateFile" Content="{x:Static Resources:SR.ManagedItemSettings_ViewCertificate}" Click="OpenCertificateFile_Click" Width="250" HorizontalAlignment="Right" />
+            <Button x:Name="ReapplyCertBindings" Content="{x:Static Resources:SR.ManagedItemSettings_ReapplyCertificate}" Click="ReapplyCertBindings_Click" Width="250" Margin="0,4,0,0" HorizontalAlignment="Right" />
+            <Label Content="{x:Static Resources:SR.ManagedItemSettings_ExportTip}" />
+            <Border>
+                <Border.Style>
+                    <Style>
+                        <Setter Property="Border.Visibility" Value="Collapsed" />
+                        <Style.Triggers>
+                            <DataTrigger Binding="{Binding SelectedItem.CertificateRevoked}" Value="False">
+                                <Setter Property="Border.Visibility" Value="Visible" />
+                            </DataTrigger>
+                        </Style.Triggers>
+                    </Style>
+                </Border.Style>
+                <Button x:Name="RevokeCertificateBtn" Content="{x:Static Resources:SR.ManagedItemSettings_RevokeCertificate}" Click="RevokeCertificateBtn_Click" Background="#FFFFC7C7" Width="250" HorizontalAlignment="Right" />
+            </Border>
+            <Border>
+                <Border.Style>
+                    <Style>
+                        <Setter Property="Border.Visibility" Value="Collapsed" />
+                        <Style.Triggers>
+                            <DataTrigger Binding="{Binding SelectedItem.CertificateRevoked}" Value="True">
+                                <Setter Property="Border.Visibility" Value="Visible" />
+                            </DataTrigger>
+                        </Style.Triggers>
+                    </Style>
+                </Border.Style>
+                <Label Content="{x:Static Resources:SR.ManagedItemSettings_CertificateRevokeWarning}" Background="#FFFFC7C7" />
+            </Border>
+        </StackPanel>
+    </DockPanel>
+</UserControl>
